+++ tests/testthat/test-allow_partial_matching.R
@@ -1,6 +1,5 @@
 
 test_that("grade_code() - allow_partial_matching works 2 errors", {
-
   expect_grade_code(
     user_code = "purrr::insistently(mean,quie = TRUE,rat = rate_backoff())",
     solution_code = "purrr::insistently(mean,quiet = TRUE,rate = rate_backoff())",